backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
pdf2image
pillow
google-generativeai
openai
python-dotenv
psutil
